python-ingesta-datos/requirements.txt
pandas
jupyter
pymssql
python-dotenv
databricks-sql-connector
pyarrow
click
pendulum
pyyaml
apscheduler